+++ performance/p2-100-million-r/index.hs
@@ -43,7 +43,7 @@ data ExampleData = MkExampleData
   { a1 :: Int64
   }
   deriving (Generic, Show, NFData)
-  deriving anyclass (ReadableFrom (Columns ExampleColumns))
+  deriving anyclass (ClickHaskell ExampleColumns)
 
 
 type ExampleColumns =